backend/requirements.txt
fastapi
uvicorn
websockets
python-multipart
openai
google-cloud-texttospeech
python-dotenv
httpx
sqlalchemy
pymysql
cryptography

